Add hotfix to Mesh with Animation loading crash
The bug was caused due to Assimp trying to import some Animation Controllers as meshes. These had out of range indices, and produced a crash when loaded.
Fixed with an extra check when creating index arrays, looking for any index out of range before saving it into the .hmesh file.